requirements.txt
matplotlib==3.3.3
nltk==3.5
numpy==1.19.4
pandas==1.2.0
plotly==4.14.1
PyYAML==5.3.1
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
tqdm==4.55.1
wordcloud==1.8.1
